Don't create invalid Bitbucket user links
Some links are created with spaces.  That's not valid, but even
more so, those pages don't exist on Bitbucket.

It seems some users only have a name but no username on Bitbucket
and some have a name and a nickname but neither are Bitbucket
usernames.

So if there's a space, only display the name but without link.
And if the display name and nickname are the same, don't add it
again.